Changes for update process

+++ ChefExtensionHandler/bin/chef-uninstall.ps1
@@ -9,33 +9,41 @@
 $chefExtensionRoot = ("{0}{1}" -f (Split-Path -Parent -Path $MyInvocation.MyCommand.Definition), "\\..")
 . $chefExtensionRoot\\bin\\shared.ps1
 
-Write-ChefStatus "uninstalling-chef" "transitioning" "Uninstalling Chef"
+if (!(Test-ChefExtensionRegistry))
+{
+  Write-ChefStatus "uninstalling-chef" "transitioning" "Uninstalling Chef"
 
-$bootstrapDirectory = "C:\\chef"
-$chefInstallDirectory = "C:\\opscode"
+  $bootstrapDirectory = "C:\\chef"
+  $chefInstallDirectory = "C:\\opscode"
 
-# uninstall does both disable and remove the service
-$result = chef-service-manager -a uninstall
-echo $result
+  # uninstall does both disable and remove the service
+  $result = chef-service-manager -a uninstall
+  echo $result
 
-# Uninstall the custom gem
-$result = gem uninstall -Ix azure-chef-extension
-echo $result
+  # Uninstall the custom gem
+  $result = gem uninstall -Ix azure-chef-extension
+  echo $result
 
-# Actual uninstall functionality
-# Get chef_pkg by matching "chef client " string with $_.Name
-$chef_pkg = Get-WmiObject -Class Win32_Product | Where-Object { $_.Name.contains("Chef Client") }
+  # Actual uninstall functionality
+  # Get chef_pkg by matching "chef client " string with $_.Name
+  $chef_pkg = Get-WmiObject -Class Win32_Product | Where-Object { $_.Name.contains("Chef Client") }
 
-# Uninstall chef_pkg
-$result = $chef_pkg.Uninstall()
-echo $result
+  # Uninstall chef_pkg
+  $result = $chef_pkg.Uninstall()
+  echo $result
 
-# clean up config files and install folder
-if (Test-Path $bootstrapDirectory) {
-  Remove-Item -Recurse -Force $bootstrapDirectory
+  # clean up config files and install folder
+  if (Test-Path $bootstrapDirectory) {
+    Remove-Item -Recurse -Force $bootstrapDirectory
+  }
+  if (Test-Path $chefInstallDirectory) {
+    Remove-Item -Recurse -Force $chefInstallDirectory
+  }
+
+  Write-ChefStatus "uninstalling-chef" "success" "Uninstalled Chef"
 }
-if (Test-Path $chefInstallDirectory) {
-  Remove-Item -Recurse -Force $chefInstallDirectory
+Else
+{
+  echo "Not tried to uninstall, as the update process is running"
+  Write-ChefStatus "updating-chef-extension" "transitioning" "Skipping Uninstall"
 }
-
-Write-ChefStatus "uninstalling-chef" "success" "Uninstalled Chef"
